Install minikube so that the getting started example works directly
The getting started snippet on the landing page requires Minikube
to be already installed.

If this is not the case, the call to kyma provision minikube will
fail.

$ kyma provision minikube
X Checking requirements
Error: Executing the 'minikube version' command with output '' and
error message 'exec: "minikube": executable file not found in $PATH'
failed

For the example to work directly, Minikube should also be installed.

Signed-off-by: Christian Berendt <berendt@23technologies.cloud>